Move calculateFileHashes to separate file
